+++ anyblok/__init__.py
@@ -9,6 +9,7 @@
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
+from pkg_resources import iter_entry_points
 from logging import getLogger
 logger = getLogger(__name__)
 
@@ -37,7 +38,6 @@ def load_init_function_from_entry_points(unittest=False):
 
 
     """
-    from pkg_resources import iter_entry_points
     for i in iter_entry_points('anyblok.init'):
         print('AnyBlok Load init: %r' % i)
         i.load()(unittest=unittest)
@@ -67,7 +67,6 @@ def configuration_post_load(unittest=False):
 
 
     """
-    from pkg_resources import iter_entry_points
     for i in iter_entry_points('anyblok_configuration.post_load'):
         logger.info('AnyBlok configuration post load: %r' % i)
         i.load()(unittest=unittest)